Rename extension class names
