+++ app/src/main/kotlin/com/absinthe/libchecker/utils/PackageUtils.kt
@@ -31,7 +31,6 @@ import com.absinthe.libchecker.constant.Constants.X86_64_STRING
 import com.absinthe.libchecker.constant.Constants.X86_STRING
 import com.absinthe.libchecker.constant.GlobalValues
 import com.absinthe.libchecker.constant.librarymap.DexLibMap
-import com.absinthe.libchecker.utils.extensions.VersionCompat
 import com.absinthe.libchecker.utils.manifest.ManifestReader
 import com.absinthe.libchecker.utils.manifest.StaticLibraryReader
 import net.dongliu.apk.parser.ApkFile
