+++ app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/CompatExtensions.kt
@@ -9,6 +9,8 @@ import kotlin.contracts.contract
 import kotlin.comparisons.reversed as kotlinReversed
 import kotlin.io.use as kotlinUse
 
+fun <T> unsafeLazy(initializer: () -> T): Lazy<T> = lazy(LazyThreadSafetyMode.NONE, initializer)
+
 // @see https://youtrack.jetbrains.com/issue/KT-35216
 @OptIn(ExperimentalContracts::class)
 inline fun <T : Closeable?, R> T.use(block: (T) -> R): R {
