+++ app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/FragmentExtensions.kt
@@ -1,11 +1,11 @@
-package com.absinthe.libchecker.utils
+package com.absinthe.libchecker.utils.extensions
 
 import android.os.Bundle
 import androidx.annotation.MainThread
 import androidx.core.os.bundleOf
+import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.Fragment
 
-fun <T> unsafeLazy(initializer: () -> T): Lazy<T> = lazy(LazyThreadSafetyMode.NONE, initializer)
 
 @MainThread
 fun <T : Fragment> T.putArguments(bundle: Bundle): T {
@@ -16,3 +16,5 @@ fun <T : Fragment> T.putArguments(bundle: Bundle): T {
 @MainThread
 fun <T : Fragment> T.putArguments(vararg pairs: Pair<String, Any?>): T =
     putArguments(bundleOf(*pairs))
+
+fun DialogFragment.isShowing() = this.dialog != null && this.dialog!!.isShowing && !this.isRemoving
