--- app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/ViewPagerKtx.kt
@@ -1,39 +1,0 @@
-package com.absinthe.libchecker.utils.extensions
-
-import android.animation.Animator
-import android.animation.TimeInterpolator
-import android.animation.ValueAnimator
-import android.view.animation.AccelerateDecelerateInterpolator
-import androidx.viewpager2.widget.ViewPager2
-
-fun ViewPager2.setCurrentItem(
-    item: Int,
-    duration: Long,
-    interpolator: TimeInterpolator = AccelerateDecelerateInterpolator(),
-    pagePxWidth: Int = width
-) {
-    val pxToDrag: Int = pagePxWidth * (item - currentItem)
-    val animator = ValueAnimator.ofInt(0, pxToDrag)
-    var previousValue = 0
-    animator.addUpdateListener { valueAnimator ->
-        val currentValue = valueAnimator.animatedValue as Int
-        val currentPxToDrag = (currentValue - previousValue).toFloat()
-        fakeDragBy(-currentPxToDrag)
-        previousValue = currentValue
-    }
-    animator.addListener(object : Animator.AnimatorListener {
-        override fun onAnimationStart(animation: Animator?) {
-            beginFakeDrag()
-        }
-
-        override fun onAnimationEnd(animation: Animator?) {
-            endFakeDrag()
-        }
-
-        override fun onAnimationCancel(animation: Animator?) {}
-        override fun onAnimationRepeat(animation: Animator?) {}
-    })
-    animator.interpolator = interpolator
-    animator.duration = duration
-    animator.start()
-}
