+++ app/src/main/kotlin/com/absinthe/libchecker/view/detail/AppBundleBottomSheetView.kt
@@ -8,7 +8,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.absinthe.libchecker.recyclerview.VerticalSpacesItemDecoration
 import com.absinthe.libchecker.recyclerview.adapter.detail.AppBundleAdapter
 import com.absinthe.libchecker.utils.extensions.dp
-import com.absinthe.libchecker.utils.unsafeLazy
+import com.absinthe.libchecker.utils.extensions.unsafeLazy
 
 class AppBundleBottomSheetView(context: Context) : LinearLayout(context) {
